+++ lib/src/pigeons/polygon_annotation_messenger.dart
@@ -41,6 +41,7 @@ class PolygonAnnotation {
     this.fillTunnelStructureColor,
     this.fillZOffset,
     this.isDraggable,
+    this.customData,
   });
 
   /// The id for annotation
@@ -89,6 +90,9 @@ class PolygonAnnotation {
   /// Property to determine whether annotation can be manually moved around map.
   bool? isDraggable;
 
+  /// JSON convertible properties associated with the annotation, used to enrich Feature GeoJSON `properties["custom_data"]` field.
+  Map<String, Object>? customData;
+
   List<Object?> _toList() {
     return <Object?>[
       id,
@@ -103,6 +107,7 @@ class PolygonAnnotation {
       fillTunnelStructureColor,
       fillZOffset,
       isDraggable,
+      customData,
     ];
   }
 
@@ -125,6 +130,8 @@ class PolygonAnnotation {
       fillTunnelStructureColor: result[9] as int?,
       fillZOffset: result[10] as double?,
       isDraggable: result[11] as bool?,
+      customData:
+          (result[12] as Map<Object?, Object?>?)?.cast<String, Object>(),
     );
   }
 
@@ -148,7 +155,8 @@ class PolygonAnnotation {
         fillPattern == other.fillPattern &&
         fillTunnelStructureColor == other.fillTunnelStructureColor &&
         fillZOffset == other.fillZOffset &&
-        isDraggable == other.isDraggable;
+        isDraggable == other.isDraggable &&
+        _deepEquals(customData, other.customData);
   }
 
   @override
@@ -169,6 +177,7 @@ class PolygonAnnotationOptions {
     this.fillTunnelStructureColor,
     this.fillZOffset,
     this.isDraggable,
+    this.customData,
   });
 
   /// The geometry that determines the location/shape of this annotation
@@ -214,6 +223,9 @@ class PolygonAnnotationOptions {
   /// Property to determine whether annotation can be manually moved around map.
   bool? isDraggable;
 
+  /// JSON convertible properties associated with the annotation, used to enrich Feature GeoJSON `properties["custom_data"]` field.
+  Map<String, Object>? customData;
+
   List<Object?> _toList() {
     return <Object?>[
       geometry,
@@ -227,6 +239,7 @@ class PolygonAnnotationOptions {
       fillTunnelStructureColor,
       fillZOffset,
       isDraggable,
+      customData,
     ];
   }
 
@@ -248,6 +261,8 @@ class PolygonAnnotationOptions {
       fillTunnelStructureColor: result[8] as int?,
       fillZOffset: result[9] as double?,
       isDraggable: result[10] as bool?,
+      customData:
+          (result[11] as Map<Object?, Object?>?)?.cast<String, Object>(),
     );
   }
 
@@ -271,7 +286,8 @@ class PolygonAnnotationOptions {
         fillPattern == other.fillPattern &&
         fillTunnelStructureColor == other.fillTunnelStructureColor &&
         fillZOffset == other.fillZOffset &&
-        isDraggable == other.isDraggable;
+        isDraggable == other.isDraggable &&
+        _deepEquals(customData, other.customData);
   }
 
   @override
@@ -342,6 +358,38 @@ class _PolygonAnnotationMessenger {
       PolygonAnnotationMessenger_PigeonCodec();
 
   final String pigeonVar_messageChannelSuffix;
+
+  Future<List<PolygonAnnotation>> getAnnotations(String managerId) async {
+    final String pigeonVar_channelName =
+        'dev.flutter.pigeon.mapbox_maps_flutter._PolygonAnnotationMessenger.getAnnotations$pigeonVar_messageChannelSuffix';
+    final BasicMessageChannel<Object?> pigeonVar_channel =
+        BasicMessageChannel<Object?>(
+      pigeonVar_channelName,
+      pigeonChannelCodec,
+      binaryMessenger: pigeonVar_binaryMessenger,
+    );
+    final Future<Object?> pigeonVar_sendFuture =
+        pigeonVar_channel.send(<Object?>[managerId]);
+    final List<Object?>? pigeonVar_replyList =
+        await pigeonVar_sendFuture as List<Object?>?;
+    if (pigeonVar_replyList == null) {
+      throw _createConnectionError(pigeonVar_channelName);
+    } else if (pigeonVar_replyList.length > 1) {
+      throw PlatformException(
+        code: pigeonVar_replyList[0]! as String,
+        message: pigeonVar_replyList[1] as String?,
+        details: pigeonVar_replyList[2],
+      );
+    } else if (pigeonVar_replyList[0] == null) {
+      throw PlatformException(
+        code: 'null-error',
+        message: 'Host platform returned null value for non-null return value.',
+      );
+    } else {
+      return (pigeonVar_replyList[0] as List<Object?>?)!
+          .cast<PolygonAnnotation>();
+    }
+  }
 
   Future<PolygonAnnotation> create(
       String managerId, PolygonAnnotationOptions annotationOption) async {