+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### 2.17.0-rc.1
+
+* [iOS] Fix annotation interaction handlers (tap, drag etc.) not working after annotation update.
+* Add the ability to get all annotations from an annotation manager.
+* Add `customData` field to annotations to be able to pass user data.
+
 ### 2.17.0-beta.1
 
   > [!NOTE]